django/drf/notes/requirements.txt
django
djangorestframework
djoser




















#pip install -r requirements.txt